Moved CSharpCodeWriter out of legacy
