+++ test/Microsoft.AspNetCore.Razor.Language.Test/CodeGeneration/TagHelperHtmlAttributeRuntimeBasicWriterTest.cs
@@ -98,7 +98,7 @@ AddHtmlAttributeValue("" "", 27, false, 28, 6, false);
         private static CSharpRenderingContext GetCSharpRenderingContext(BasicWriter writer)
         {
             var options = RazorCodeGenerationOptions.CreateDefault();
-            var codeWriter = new Legacy.CSharpCodeWriter();
+            var codeWriter = new CSharpCodeWriter();
             var context = new CSharpRenderingContext()
             {
                 Writer = codeWriter,
